BigodeBot/requirements.txt
# Requirements for BigodeBot - DayZ Server Management System
# Version: 2.0
# Python: 3.10+

# ==========================================
# CORE FRAMEWORK
# ==========================================
Flask==3.0.0
Flask-Session==0.5.0
discord.py==2.3.2
aiohttp==3.9.1

# ==========================================
# HTTP & NETWORKING
# ==========================================
requests==2.31.0
python-dotenv==1.0.0

# ==========================================
# IMAGE PROCESSING & VISUALIZATION
# ==========================================
Pillow==10.1.0
matplotlib==3.8.2

# ==========================================
# UTILITIES
# ==========================================
pytz==2023.3

# ==========================================
# AI INTEGRATION
# ==========================================
google-genai==0.3.0

# ==========================================
# VALIDATION & DATA MODELS (NEW)
# ==========================================
pydantic==2.5.3
pydantic-settings==2.1.0

# ==========================================
# TESTING & QUALITY (NEW)
# ==========================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# ==========================================
# CODE FORMATTING & LINTING (NEW)
# ==========================================
black==23.12.1
ruff==0.1.9
mypy==1.7.1

# ==========================================
# SECURITY
# ==========================================
bandit==1.7.6
defusedxml==0.7.1

# ==========================================
# PRODUCTION SERVERS
# ==========================================
gunicorn==21.2.0
waitress==2.1.2

# ==========================================
# DEVELOPMENT ONLY (install with: pip install -r requirements-dev.txt)
# ==========================================
# ipython==8.18.1
# ipdb==0.13.13
